requirements.txt
streamlit
Pillow
pandas
requests
numpy
matplotlib